fix of missing icon bug
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4203 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/font_handler.cc
@@ -899,7 +899,7 @@ void Font_Handler::get_size
 		int32_t line_w, line_h;
 		TTF_SizeUTF8(&font, line.c_str(), &line_w, &line_h);
 
-		if (w < line_w)
+		if (static_cast<int32_t>(w) < line_w)
 			w = line_w;
 		h += line_h;
 	}
